requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-filter>=24.0
django-cors-headers>=4.3
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
requests>=2.31
python-decouple>=3.8
whitenoise>=6.6
gunicorn>=21.2
dj-database-url>=2.1
